requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
mysql-connector-python
python-dotenv
firebase-admin
sentence-transformers
spacy
numpy
tensorflow
twilio
Flask-Migrate
gunicorn
psycopg2-binary